feat: VideoCard added with custom logic hook
