+++ src/components/VideoCard/VideoCard.jsx
@@ -6,37 +6,20 @@ import {
   BsFillHandThumbsUpFill,
   BsFillHandThumbsDownFill,
 } from "react-icons/bs";
-import { useNavigate } from "react-router-dom";
 import { BadgeButton } from "../Buttons/BadgeButton";
-import { useVideoLike } from "../../hooks/useVideoLike";
-import { useDataProvider } from "../../context/data-context";
-import { ACTION_TYPE, openPlaylist } from "../../utils";
-import { useWatchLater } from "../../hooks/useWatchLater";
-import { useAuth } from "../../hooks/useAuth";
+import { useVideoCard } from "./useVideoCard";
 
 export function VideoCard({ video }) {
-  const { _id, title, creator, categoryName, thumbnail } = video;
-  const { isAuthenticated } = useAuth();
+  const { title, creator, categoryName, thumbnail } = video;
 
-  const navigate = useNavigate();
-  const videoCardClickHandler = () => {
-    navigate(`/videos/${_id}`);
-  };
-
-  const { dispatch } = useDataProvider();
-  const { likedVideo, likeHandler } = useVideoLike(video);
-  const { inWatchLater, watchLaterHandler } = useWatchLater(video);
-
-  const playlistHandler = () => {
-    if (isAuthenticated) {
-      openPlaylist(video, dispatch);
-    } else {
-      dispatch({
-        type: ACTION_TYPE.ACTIVATE_ALERT,
-        payload: { message: "You need to login", color: "green" },
-      });
-    }
-  };
+  const {
+    videoCardClickHandler,
+    likedVideo,
+    likeHandler,
+    playlistHandler,
+    inWatchLater,
+    watchLaterHandler,
+  } = useVideoCard(video);
 
   return (
     <div className="card__flexcolumn card__flexcolumn--lg card__shadow--green video--card">
